+++ lib/src/bson_codec.dart
@@ -3,9 +3,10 @@ import 'types/base/bson_object.dart';
 import 'types/bson_map.dart';
 
 class BsonCodec {
+  /// Serializes a document into a BSON binary object
   static BsonBinary serialize(var object, [int offset = 0]) {
     if (!((object is Map) || (object is List))) {
-      throw Exception('Invalid value for BSON serialize: $object');
+      throw Exception('Invalid value for BsonCodec.serialize: $object');
     }
     var bsonObject = BsonObject.from(object, bsonSerialization);
     var buffer = BsonBinary(bsonObject.totalByteLength + offset);
@@ -15,6 +16,7 @@ class BsonCodec {
     return buffer;
   }
 
+  /// Converts BSON binary object into a document.
   static Map<String, dynamic> deserialize(BsonBinary buffer) {
     buffer.offset = 0;
     if (buffer.byteList.length < 5) {