+++ lib/bson.dart
@@ -5,31 +5,15 @@ export 'src/classes/js_code.dart';
 export 'src/classes/object_id.dart';
 export 'src/classes/timestamp.dart';
 
-//export 'src/types/bson_decimal_128.dart';
-//export 'src/types/bson_map.dart';
-//export 'src/types/bson_array.dart';
-//export 'src/types/bson_double.dart';
-//export 'src/types/bson_string.dart';
-//export 'src/types/bson_null.dart';
-//export 'src/types/bson_boolean.dart';
-//export 'src/types/bson_regexp.dart';
-//export 'src/types/bson_db_ref.dart';
 export 'src/types/dbpointer.dart';
-//export 'src/types/bson_int.dart';
-//export 'src/types/bson_long.dart';
-//export 'src/types/bson_timestamp.dart';
 export 'src/types/base/serialization_parameters.dart';
-//export 'src/types/bson_date.dart';
-//export 'src/types/bson_object_id.dart';
 export 'src/types/bson_binary.dart';
-//export 'src/types/bson_code.dart';
 
 export 'src/bson_codec.dart';
 export 'src/codec.dart';
 export 'src/ejson_codec.dart';
 export 'src/object_codec.dart';
 
-export 'src/utils/statics.dart';
 export 'src/utils/types_def.dart';
 
 export 'src/object_serialization/bson_serializable_mixin.dart';