+++ lib/src/classes/js_code.dart
@@ -1,5 +1,4 @@
-// ignore_for_file: no_leading_underscores_for_local_identifiers
-
+/// This class is used to store Javascript code
 class JsCode {
   JsCode(this.code);