+++ lib/src/object_serialization/bson_serializable_mixin.dart
@@ -1,4 +1,5 @@
-import 'package:bson/bson.dart';
+import '../object_codec.dart';
+import '../types/bson_binary.dart';
 
 mixin BsonSerializable {
   Map<String, dynamic> get toBson;