+++ lib/src/types/bson_binary.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 
 import 'package:fixnum/fixnum.dart';
 
-import '../../bson.dart';
+import '../utils/types_def.dart';
 import 'base/bson_object.dart';
 import 'bson_uuid.dart';