+++ lib/src/types/bson_long.dart
@@ -1,7 +1,8 @@
 import 'package:fixnum/fixnum.dart';
 
-import '../../bson.dart';
+import '../utils/types_def.dart';
 import 'base/bson_object.dart';
+import 'bson_binary.dart';
 
 class BsonLong extends BsonObject {
   BsonLong(this.data);