+++ lib/src/types/bson_null.dart
@@ -1,5 +1,5 @@
-import '../../bson.dart';
 import 'base/bson_object.dart';
+import 'bson_binary.dart';
 
 class BsonNull extends BsonObject {
   BsonNull();