+++ lib/src/types/dbpointer.dart
@@ -1,7 +1,6 @@
-import 'package:bson/src/types/base/bson_object.dart';
-
 import '../classes/object_id.dart';
 import '../utils/types_def.dart';
+import 'base/bson_object.dart';
 import 'bson_binary.dart';
 import 'bson_object_id.dart';
 import 'bson_string.dart';