+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 This release contains many breaking changes. Here is a list of the most noticeable:
 
+- `BSON` has been deprecated in favor of `BsonCodec`
 - In general the logic is changed so that the document to be encoded should contain normal "Dart object" and the document returned will do the same. In input the system will continue to accept BsonObjects anyway, the issue can be with the returned Objects. At present the more important exceptions are BsonBinary and DbPointer.
 - To address the above logic some new classes have been created: JsCode and DbRef.
 - The DbRef values were stored as DbPointers. Now they are stored correctly as a particular type of Map (contains $ref and $id elements). Beware that, if you have stored Dbref values, now they will be returnde as DbPointers.
@@ -11,7 +12,7 @@ Also an optional $db value is accepted and managed from the new DbRef class.
 - The JsonCode class is used to clearly separate normal Strings from Javascript code.
 - The DbPointer class is deprecated (as per Bson specifications).
 - dataSize() and byteLenght() have been deprecated in favor of contentLength and totalByteLength getters.
-- the Bson clases now are intended to be used internally, so they are no more exported explicitly in the bson package.
+- the Bson classes now are intended to be used internally, so they are no more exported explicitly in the bson package.
 
 5.0.0-5.0.beta