+++ CHANGELOG.md
@@ -9,12 +9,18 @@ This release contains many breaking changes. Here is a list of the most noticeab
 - To address the above logic some new classes have been created: JsCode and DbRef.
 - The DbRef values were stored as DbPointers. Now they are stored correctly as a particular type of Map (contains $ref and $id elements). Beware that, if you have stored Dbref values, now they will be returnde as DbPointers.
 Also an optional $db value is accepted and managed from the new DbRef class.
-- The JsonCode class is used to clearly separate normal Strings from Javascript code.
+- The JsCode class is used to clearly separate normal Strings from Javascript code.
+- The RegExp class is used instead of the BsonRegexp one.
 - The DbPointer class is deprecated (as per Bson specifications).
 - dataSize() and byteLenght() have been deprecated in favor of contentLength and totalByteLength getters.
 - the Bson classes now are intended to be used internally, so they are no more exported explicitly in the bson package.
 
-5.0.0-5.0.beta
+## 5.0.0-6.0.beta
+
+- General Clean-up
+- Deprecated `ObjectId.toHexString()` in favor of the `ObjectId.$oid` getter
+
+## 5.0.0-5.0.beta
 
 - Added `EJson.stringify()` method.
 - Added `EJson.parse()` method.